# Add benchmarks for serializer performance.