--- a/FacilityCSharp.sln
+++ b/FacilityCSharp.sln
@@ -46,6 +46,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FacilityConformance", "src\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XmlDocGen", "tools\XmlDocGen\XmlDocGen.csproj", "{636E5B36-4AA8-47D8-884C-AE4D80B49A14}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Facility.Benchmarks", "tests\Facility.Benchmarks\Facility.Benchmarks.csproj", "{C8975C56-9329-452D-8FBB-8252F6755111}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -100,6 +102,10 @@ Global
 		{636E5B36-4AA8-47D8-884C-AE4D80B49A14}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{636E5B36-4AA8-47D8-884C-AE4D80B49A14}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{636E5B36-4AA8-47D8-884C-AE4D80B49A14}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C8975C56-9329-452D-8FBB-8252F6755111}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C8975C56-9329-452D-8FBB-8252F6755111}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C8975C56-9329-452D-8FBB-8252F6755111}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C8975C56-9329-452D-8FBB-8252F6755111}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE