--- a/tests/Facility.Benchmarks/BenchmarkService.fsd
+++ b/tests/Facility.Benchmarks/BenchmarkService.fsd
@@ -1,0 +1,26 @@
+/// Common service elements.
+[csharp(namespace: Facility.Benchmarks)]
+service BenchmarkService
+{
+	[http(method: GET, path: "/users")]
+	method getUsers
+	{
+		limit: int32;
+	}:
+	{
+		items: User[];
+	}
+
+	data User
+	{
+		id: int32;
+
+		name: string;
+
+		email: string;
+
+		phone: string;
+
+		about: string;
+	}
+}