--- a/tests/Facility.Benchmarks/GlobalSuppressions.cs
+++ b/tests/Facility.Benchmarks/GlobalSuppressions.cs
@@ -1,0 +1,3 @@
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Reliability", "CA2007:Consider calling ConfigureAwait on the awaited task", Justification = "Not needed.")]