--- a/tests/Facility.Benchmarks/Program.cs
+++ b/tests/Facility.Benchmarks/Program.cs
@@ -1,0 +1,4 @@
+using BenchmarkDotNet.Running;
+using Facility.Benchmarks;
+
+BenchmarkRunner.Run<HttpBenchmarks>();