--- a/tests/Facility.Benchmarks/SerializerInfo.cs
+++ b/tests/Facility.Benchmarks/SerializerInfo.cs
@@ -1,0 +1,8 @@
+using Facility.Core;
+
+namespace Facility.Benchmarks;
+
+public record SerializerInfo(ServiceSerializer ServiceSerializer)
+{
+	public override string ToString() => ServiceSerializer.GetType().Name[0..^"ServiceSerializer".Length];
+}