… tests/Facility.Benchmarks/BenchmarkServiceMethods.g.cs
… tests/Facility.Benchmarks/DelegatingBenchmarkService.g.cs
… tests/Facility.Benchmarks/GetUsersRequestDto.g.cs
… tests/Facility.Benchmarks/GetUsersResponseDto.g.cs
… tests/Facility.Benchmarks/Http/BenchmarkServiceHttpHandler.g.cs
… tests/Facility.Benchmarks/Http/BenchmarkServiceHttpMapping.g.cs
… tests/Facility.Benchmarks/Http/HttpClientBenchmarkService.g.cs
… tests/Facility.Benchmarks/IBenchmarkService.g.cs
… tests/Facility.Benchmarks/UserDto.g.cs