```
|   Method | UserCount |     Serializer |     Mean |     Error |    StdDev |   Median |    Gen 0 |    Gen 1 | Allocated |
|--------- |---------- |--------------- |---------:|----------:|----------:|---------:|---------:|---------:|----------:|
| GetUsers |         1 | NewtonsoftJson | 2.164 ms | 0.0076 ms | 0.0224 ms | 2.162 ms | 351.5625 | 121.0938 |  4,503 KB |
| GetUsers |         1 | SystemTextJson | 1.072 ms | 0.0056 ms | 0.0163 ms | 1.068 ms |  64.4531 |  25.3906 |    823 KB |
| GetUsers |      1000 | NewtonsoftJson | 3.505 ms | 0.0046 ms | 0.0130 ms | 3.505 ms | 382.8125 | 191.4063 |  4,921 KB |
| GetUsers |      1000 | SystemTextJson | 1.890 ms | 0.0038 ms | 0.0104 ms | 1.889 ms |  89.8438 |  42.9688 |  1,168 KB |
```